+++ tests/testthat/test-order.R
@@ -1243,65 +1243,65 @@ test_that("can sort empty data frames (#356)", {
 })
 
 # ------------------------------------------------------------------------------
-# vec_order_base
+# vec_order
 
 test_that("can request NAs sorted first", {
-  expect_equal(vec_order_base(c(1, NA), "asc", "largest"), 1:2)
-  expect_equal(vec_order_base(c(1, NA), "desc", "largest"), 2:1)
+  expect_equal(vec_order(c(1, NA), "asc", "largest"), 1:2)
+  expect_equal(vec_order(c(1, NA), "desc", "largest"), 2:1)
 
-  expect_equal(vec_order_base(c(1, NA), "asc", "smallest"), 2:1)
-  expect_equal(vec_order_base(c(1, NA), "desc", "smallest"), 1:2)
+  expect_equal(vec_order(c(1, NA), "asc", "smallest"), 2:1)
+  expect_equal(vec_order(c(1, NA), "desc", "smallest"), 1:2)
 })
 
 test_that("can order complex vectors", {
   x <- complex(real = c(1, 2, 2, 3, 3), imaginary = c(5, 4, 3, 2, NA))
 
-  expect_equal(vec_order_base(x, direction = "asc", na_value = "largest"), c(1, 3, 2, 4, 5))
-  expect_equal(vec_order_base(x, direction = "desc", na_value = "largest"), rev(c(1, 3, 2, 4, 5)))
-  expect_equal(vec_order_base(x, direction = "asc", na_value = "smallest"), c(5, 1, 3, 2, 4))
-  expect_equal(vec_order_base(x, direction = "desc", na_value = "smallest"), rev(c(5, 1, 3, 2, 4)))
+  expect_equal(vec_order(x, direction = "asc", na_value = "largest"), c(1, 3, 2, 4, 5))
+  expect_equal(vec_order(x, direction = "desc", na_value = "largest"), rev(c(1, 3, 2, 4, 5)))
+  expect_equal(vec_order(x, direction = "asc", na_value = "smallest"), c(5, 1, 3, 2, 4))
+  expect_equal(vec_order(x, direction = "desc", na_value = "smallest"), rev(c(5, 1, 3, 2, 4)))
 })
 
 test_that("can order tibbles that contain non-comparable objects", {
-  expect_equal(vec_order_base(data_frame(x = list(10, 2, 1))), 1:3)
+  expect_equal(vec_order(data_frame(x = list(10, 2, 1))), 1:3)
 })
 
 test_that("can order matrices and arrays (#306)", {
   x <- matrix(c(1, 1, 1, 1, 2, 1), ncol = 2)
-  expect_identical(vec_order_base(x), c(1L, 3L, 2L))
+  expect_identical(vec_order(x), c(1L, 3L, 2L))
 
   x <- array(1:8, c(2, 2, 2))
   x[2] <- 1
   x[3] <- 5
-  expect_identical(vec_order_base(x), 2:1)
+  expect_identical(vec_order(x), 2:1)
 })
 
 test_that("can order zero column data frames (#356, #1499)", {
   df <- data_frame()
-  expect_identical(vec_order_base(df), integer())
+  expect_identical(vec_order(df), integer())
 
   df <- data_frame(.size = 5L)
-  expect_identical(vec_order_base(df), 1:5)
+  expect_identical(vec_order(df), 1:5)
 
   df <- data_frame(.size = 5L)
-  expect_identical(vec_order_base(df, direction = "desc"), 1:5)
+  expect_identical(vec_order(df, direction = "desc"), 1:5)
 })
 
 test_that("can order zero row data frames (#356, #1499)", {
   df <- data.frame(x = numeric())
-  expect_identical(vec_order_base(df), integer())
+  expect_identical(vec_order(df), integer())
 
   df <- data.frame(x = numeric(), y = integer())
-  expect_identical(vec_order_base(df), integer())
+  expect_identical(vec_order(df), integer())
 
   df <- data.frame(x = numeric(), y = integer())
-  expect_identical(vec_order_base(df, direction = "desc"), integer())
+  expect_identical(vec_order(df, direction = "desc"), integer())
 })
 
 test_that("can order data frames with data frame columns (#527)", {
   expect_equal(
-    vec_order_base(iris),
-    vec_order_base(data_frame(iris = iris))
+    vec_order(iris),
+    vec_order(data_frame(iris = iris))
   )
 })
 
@@ -1309,16 +1309,16 @@ test_that("can order data frames (and subclasses) with matrix columns", {
   df <- new_data_frame(n = 2L)
 
   df$x <- new_data_frame(list(y = matrix(1:2, 2)))
-  expect_identical(vec_order_base(df), 1:2)
+  expect_identical(vec_order(df), 1:2)
 
   df$x <- tibble::tibble(y = matrix(1:2, 2))
-  expect_identical(vec_order_base(df), 1:2)
+  expect_identical(vec_order(df), 1:2)
 })
 
 test_that("classed proxies do not affect performance (tidyverse/dplyr#5423)", {
   skip_on_cran()
   x <- glue::glue("{1:10000}")
-  expect_time_lt(vec_order_base(x), 0.2)
+  expect_time_lt(vec_order(x), 0.2)
 })
 
 test_that("can order data frames that don't allow removing the column names (#1298)", {
@@ -1335,11 +1335,11 @@ test_that("can order data frames that don't allow removing the column names (#12
 
   df <- foobar(data.frame(x = 1, y = 2))
 
-  expect_silent(expect_identical(vec_order_base(df), 1L))
+  expect_silent(expect_identical(vec_order(df), 1L))
 })
 
 test_that("missing values in lists are respected (#1401)", {
   x <- list(1, NULL, 2, NULL)
-  expect_identical(vec_order_base(x, na_value = "largest"), c(1L, 3L, 2L, 4L))
-  expect_identical(vec_order_base(x, na_value = "smallest"), c(2L, 4L, 1L, 3L))
+  expect_identical(vec_order(x, na_value = "largest"), c(1L, 3L, 2L, 4L))
+  expect_identical(vec_order(x, na_value = "smallest"), c(2L, 4L, 1L, 3L))
 })